app/requirements.txt
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
fastapi==0.111.0
uvicorn==0.30.6
boto3==1.34.0
chromadb==1.5.7
sentence-transformers==3.0.1
pypdf==4.3.1
python-dotenv==1.0.1
httpx==0.27.0
requests==2.32.3
